Displaying but not in correct position

+++ Colony/Node.cs
@@ -11,11 +11,11 @@ internal class Node {
 	/// <summary>
 	/// X coordinate
 	/// </summary>
-	public int X { get; set; }
+	public double X { get; set; }
 	/// <summary>
 	/// Y coordinate
 	/// </summary>
-	public int Y { get; set; }
+	public double Y { get; set; }
 	/// <summary>
 	/// IF this cell is marked as discovered
 	/// </summary>
